requirements.txt
PGPy~=0.6.0
discord.py~=2.2.2
discord.py-self~=2.0.0